# Use Data.Map.Strict