--- a/compiler/repl/Repl/Actions/TypeSearch.hs
+++ b/compiler/repl/Repl/Actions/TypeSearch.hs
@@ -6,7 +6,7 @@ module Repl.Actions.TypeSearch
 where
 
 import Data.Foldable
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import qualified Data.Text as T
 import qualified Language.Mimsa.Actions.Monad as Actions
 import qualified Language.Mimsa.Actions.Typecheck as Actions