--- a/compiler/repl/Repl/Persistence.hs
+++ b/compiler/repl/Repl/Persistence.hs
@@ -17,7 +17,7 @@ import Control.Monad.Reader
 import qualified Data.Aeson as JSON
 import qualified Data.ByteString.Lazy as LBS
 import Data.Functor
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import qualified Data.Text as T