--- a/compiler/repl/ReplNew/Actions/ListModules.hs
+++ b/compiler/repl/ReplNew/Actions/ListModules.hs
@@ -7,7 +7,7 @@ module ReplNew.Actions.ListModules
 where
 
 import Data.Foldable (traverse_)
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Text (Text)
 import qualified Language.Mimsa.Actions.Helpers.LookupExpression as Actions
 import qualified Language.Mimsa.Actions.Modules.Typecheck as Actions