--- a/compiler/server/Server/Endpoints/Expression.hs
+++ b/compiler/server/Server/Endpoints/Expression.hs
@@ -15,8 +15,8 @@ where
 
 import qualified Data.Aeson as JSON
 import Data.Bifunctor
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.OpenApi hiding (Server)
 import qualified Data.Text as T
 import GHC.Generics