--- a/compiler/server/Server/Handlers.hs
+++ b/compiler/server/Server/Handlers.hs
@@ -33,8 +33,8 @@ import Data.Coerce
 import Data.Foldable (traverse_)
 import Data.Functor
 import qualified Data.List.NonEmpty as NE
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.OpenApi (ToSchema)
 import Data.Set (Set)
 import qualified Data.Set as S