--- a/compiler/src/Language/Mimsa/Actions/Compile.hs
+++ b/compiler/src/Language/Mimsa/Actions/Compile.hs
@@ -13,8 +13,8 @@ import Control.Monad.Except
 import Data.Bifunctor (first)
 import Data.Coerce
 import Data.Foldable (traverse_)
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import qualified Data.Text as T