--- a/compiler/src/Language/Mimsa/Actions/Graph.hs
+++ b/compiler/src/Language/Mimsa/Actions/Graph.hs
@@ -5,8 +5,8 @@ module Language.Mimsa.Actions.Graph (graphExpression, graphProject) where
 
 import Control.Monad.Except
 import Data.Bifunctor (first)
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Language.Mimsa.Actions.Monad as Actions
 import Language.Mimsa.Project
 import Language.Mimsa.Store.DepGraph