--- a/compiler/src/Language/Mimsa/Actions/Helpers/Build.hs
+++ b/compiler/src/Language/Mimsa/Actions/Helpers/Build.hs
@@ -4,8 +4,8 @@
 module Language.Mimsa.Actions.Helpers.Build (doJobs, getMissing, Plan (..), State (..), Job, Inputs) where
 
 import Control.Parallel.Strategies
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.Printer