--- a/compiler/src/Language/Mimsa/Actions/Modules/Bind.hs
+++ b/compiler/src/Language/Mimsa/Actions/Modules/Bind.hs
@@ -6,7 +6,7 @@ module Language.Mimsa.Actions.Modules.Bind
   )
 where
 
-import Data.Map (Map)
+import Data.Map.Strict (Map)
 import Data.Text (Text)
 import qualified Language.Mimsa.Actions.Modules.RunTests as Actions
 import qualified Language.Mimsa.Actions.Modules.Typecheck as Actions