--- a/compiler/src/Language/Mimsa/Actions/Modules/Check.hs
+++ b/compiler/src/Language/Mimsa/Actions/Modules/Check.hs
@@ -3,7 +3,7 @@
 
 module Language.Mimsa.Actions.Modules.Check (checkModule) where
 
-import Data.Map (Map)
+import Data.Map.Strict (Map)
 import Data.Text (Text)
 import qualified Language.Mimsa.Actions.Modules.RunTests as Actions
 import qualified Language.Mimsa.Actions.Monad as Actions