--- a/compiler/src/Language/Mimsa/Actions/Modules/Typecheck.hs
+++ b/compiler/src/Language/Mimsa/Actions/Modules/Typecheck.hs
@@ -5,8 +5,8 @@ module Language.Mimsa.Actions.Modules.Typecheck
 where
 
 import Control.Monad.Except
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Text (Text)
 import qualified Language.Mimsa.Actions.Monad as Actions
 import Language.Mimsa.Modules.HashModule