--- a/compiler/src/Language/Mimsa/Actions/Monad.hs
+++ b/compiler/src/Language/Mimsa/Actions/Monad.hs
@@ -28,8 +28,8 @@ where
 
 import Control.Monad.Except
 import Control.Monad.State
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text (Text)