--- a/compiler/src/Language/Mimsa/Actions/Optimise.hs
+++ b/compiler/src/Language/Mimsa/Actions/Optimise.hs
@@ -10,8 +10,8 @@ where
 
 import Control.Monad.Except
 import Data.Bifunctor
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import qualified Language.Mimsa.Actions.Helpers.Build as Build
 import qualified Language.Mimsa.Actions.Helpers.CheckStoreExpression as Actions