--- a/compiler/src/Language/Mimsa/Actions/Typecheck.hs
+++ b/compiler/src/Language/Mimsa/Actions/Typecheck.hs
@@ -12,8 +12,8 @@ where
 import Control.Monad.Except
 import Data.Bifunctor (first)
 import Data.Foldable (traverse_)
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Language.Mimsa.Actions.Helpers.Build as Build