--- a/compiler/src/Language/Mimsa/Actions/Upgrade.hs
+++ b/compiler/src/Language/Mimsa/Actions/Upgrade.hs
@@ -6,8 +6,8 @@ module Language.Mimsa.Actions.Upgrade (upgradeByName) where
 import Control.Applicative
 import Control.Monad.Except
 import Data.Bifunctor (first)
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Maybe
 import Data.Set (Set)
 import qualified Data.Set as S