--- a/compiler/src/Language/Mimsa/Backend/Javascript/Printer.hs
+++ b/compiler/src/Language/Mimsa/Backend/Javascript/Printer.hs
@@ -10,7 +10,7 @@ module Language.Mimsa.Backend.Javascript.Printer
   )
 where
 
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text (Text)