--- a/compiler/src/Language/Mimsa/Backend/Output.hs
+++ b/compiler/src/Language/Mimsa/Backend/Output.hs
@@ -18,8 +18,8 @@ import Data.Bifunctor
 import Data.Coerce
 import Data.Functor
 import Data.List (intersperse)
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Data.Text as T