--- a/compiler/src/Language/Mimsa/Backend/Typescript/Monad.hs
+++ b/compiler/src/Language/Mimsa/Backend/Typescript/Monad.hs
@@ -26,8 +26,8 @@ import Control.Monad.Writer
 import Data.Either
 import Data.List.NonEmpty (NonEmpty ((:|)))
 import qualified Data.List.NonEmpty as NE
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.Backend.Typescript.Types