--- a/compiler/src/Language/Mimsa/Codegen.hs
+++ b/compiler/src/Language/Mimsa/Codegen.hs
@@ -19,8 +19,8 @@ where
 import qualified Data.Aeson as JSON
 import Data.Either (isRight)
 import Data.Functor
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.OpenApi (ToSchema)
 import GHC.Generics
 import Language.Mimsa.Codegen.ApplicativeApply