--- a/compiler/src/Language/Mimsa/Interpreter/PatternMatch.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/PatternMatch.hs
@@ -7,7 +7,7 @@ module Language.Mimsa.Interpreter.PatternMatch
 where
 
 import Control.Monad.Except
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Maybe (fromMaybe)
 import Data.Monoid
 import qualified Data.Set as S