--- a/compiler/src/Language/Mimsa/Interpreter/RecordAccess.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/RecordAccess.hs
@@ -1,7 +1,7 @@
 module Language.Mimsa.Interpreter.RecordAccess (interpretRecordAccess) where
 
 import Control.Monad.Except
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Language.Mimsa.Interpreter.Types
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error.InterpreterError