--- a/compiler/src/Language/Mimsa/Modules/Check.hs
+++ b/compiler/src/Language/Mimsa/Modules/Check.hs
@@ -10,8 +10,8 @@ module Language.Mimsa.Modules.Check
   )
 where
 
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import Language.Mimsa.Typechecker.Elaborate
 import Language.Mimsa.Types.AST