--- a/compiler/src/Language/Mimsa/Modules/Dependencies.hs
+++ b/compiler/src/Language/Mimsa/Modules/Dependencies.hs
@@ -14,8 +14,8 @@ where
 -- work out the dependencies between definitions inside a module
 
 import Control.Monad.Except
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Maybe
 import Data.Monoid (First (..))
 import Data.Set (Set)