--- a/compiler/src/Language/Mimsa/Modules/Monad.hs
+++ b/compiler/src/Language/Mimsa/Modules/Monad.hs
@@ -15,8 +15,8 @@ where
 import Control.Monad.Except
 import Data.Coerce
 import Data.Foldable
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error