--- a/compiler/src/Language/Mimsa/Modules/Pretty.hs
+++ b/compiler/src/Language/Mimsa/Modules/Pretty.hs
@@ -2,8 +2,8 @@
 
 module Language.Mimsa.Modules.Pretty (modulePretty) where
 
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import Language.Mimsa.Printer
 import Language.Mimsa.Typechecker.Elaborate