--- a/compiler/src/Language/Mimsa/Modules/ToStoreExprs.hs
+++ b/compiler/src/Language/Mimsa/Modules/ToStoreExprs.hs
@@ -11,8 +11,8 @@ import Control.Monad.Except
 import Control.Monad.Identity
 import Data.Bifunctor
 import Data.Functor (($>))
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Maybe
 import Data.Set (Set)
 import qualified Data.Set as S