--- a/compiler/src/Language/Mimsa/Modules/Typecheck.hs
+++ b/compiler/src/Language/Mimsa/Modules/Typecheck.hs
@@ -10,8 +10,8 @@ import Control.Monad.Except
 import Data.Bifunctor
 import Data.Coerce
 import Data.Foldable
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Maybe
 import Data.Set (Set)
 import qualified Data.Set as S