--- a/compiler/src/Language/Mimsa/Modules/Uses.hs
+++ b/compiler/src/Language/Mimsa/Modules/Uses.hs
@@ -7,7 +7,7 @@ module Language.Mimsa.Modules.Uses
   )
 where
 
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import qualified Language.Mimsa.TypeUtils as MT