--- a/compiler/src/Language/Mimsa/Printer.hs
+++ b/compiler/src/Language/Mimsa/Printer.hs
@@ -11,8 +11,8 @@ where
 -- prettyPrint is for debug output
 -- prettyDoc returns a Prettyprinter doc for nicer output
 
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text (Text)