--- a/compiler/src/Language/Mimsa/Project/TypeSearch.hs
+++ b/compiler/src/Language/Mimsa/Project/TypeSearch.hs
@@ -13,8 +13,8 @@ import Data.Bifunctor (first)
 import Data.Either (isRight)
 import Data.Functor
 import qualified Data.List.NonEmpty as NE
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Text (Text)
 import Language.Mimsa.Parser.Language
 import Language.Mimsa.Typechecker.NormaliseTypes