--- a/compiler/src/Language/Mimsa/Project/Usages.hs
+++ b/compiler/src/Language/Mimsa/Project/Usages.hs
@@ -3,8 +3,8 @@
 module Language.Mimsa.Project.Usages where
 
 import Data.Bifunctor
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.Project.Helpers (getCurrentBindings)