--- a/compiler/src/Language/Mimsa/Project/Versions.hs
+++ b/compiler/src/Language/Mimsa/Project/Versions.hs
@@ -10,7 +10,7 @@ import Control.Monad.Except
 import Data.Bifunctor (first)
 import Data.List.NonEmpty (NonEmpty)
 import qualified Data.List.NonEmpty as NE
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.Project