--- a/compiler/src/Language/Mimsa/Store/ExtractVars.hs
+++ b/compiler/src/Language/Mimsa/Store/ExtractVars.hs
@@ -3,7 +3,7 @@ module Language.Mimsa.Store.ExtractVars
   )
 where
 
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.Types.AST