--- a/compiler/src/Language/Mimsa/Store/Helpers.hs
+++ b/compiler/src/Language/Mimsa/Store/Helpers.hs
@@ -3,7 +3,7 @@
 module Language.Mimsa.Store.Helpers where
 
 import Control.Monad.Except
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Language.Mimsa.Types.Error.StoreError
 import Language.Mimsa.Types.Store