--- a/compiler/src/Language/Mimsa/Store/Persistence.hs
+++ b/compiler/src/Language/Mimsa/Store/Persistence.hs
@@ -12,8 +12,8 @@ where
 
 import Control.Monad.Except
 import Control.Monad.Logger
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.Project.Helpers