--- a/compiler/src/Language/Mimsa/Store/Resolver.hs
+++ b/compiler/src/Language/Mimsa/Store/Resolver.hs
@@ -6,8 +6,8 @@ module Language.Mimsa.Store.Resolver
 where
 
 import Data.Coerce
-import Data.Map (Map)
-import qualified Data.Map as M
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as M
 import Data.Maybe (catMaybes, isJust)
 import qualified Data.Set as S
 import Language.Mimsa.Store.ExtractTypes (extractTypes)