--- a/compiler/src/Language/Mimsa/Transform/FindUnused.hs
+++ b/compiler/src/Language/Mimsa/Transform/FindUnused.hs
@@ -1,6 +1,6 @@
 module Language.Mimsa.Transform.FindUnused (findUnused, removeBindings, removeUnused) where
 
-import qualified Data.Map as M
+import qualified Data.Map.Strict as M
 import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.ExprUtils