… compiler/src/Language/Mimsa/Tests/Generate.hs
… compiler/src/Language/Mimsa/Tests/Test.hs
… compiler/src/Language/Mimsa/Transform/Inliner.hs
… compiler/src/Language/Mimsa/Transform/TrimDeps.hs
… compiler/src/Language/Mimsa/TypeUtils.hs
… compiler/src/Language/Mimsa/Typechecker/BuiltIns.hs
… compiler/src/Language/Mimsa/Typechecker/CreateEnv.hs
… compiler/src/Language/Mimsa/Typechecker/DataTypes.hs
… compiler/src/Language/Mimsa/Typechecker/Elaborate.hs
… compiler/src/Language/Mimsa/Typechecker/Environment.hs
… compiler/src/Language/Mimsa/Typechecker/Exhaustiveness.hs
… compiler/src/Language/Mimsa/Typechecker/Generalise.hs
… compiler/src/Language/Mimsa/Typechecker/NormaliseTypes.hs
… compiler/src/Language/Mimsa/Typechecker/NumberVars.hs
… compiler/src/Language/Mimsa/Typechecker/ScopeTypeVar.hs
… compiler/src/Language/Mimsa/Typechecker/TcMonad.hs
… compiler/src/Language/Mimsa/Typechecker/Typecheck.hs
… compiler/src/Language/Mimsa/Typechecker/TypedHoles.hs
… compiler/src/Language/Mimsa/Typechecker/Unify.hs
… compiler/src/Language/Mimsa/Types/AST/DataType.hs
… compiler/src/Language/Mimsa/Types/AST/Expr.hs
… compiler/src/Language/Mimsa/Types/AST/Pattern.hs
… compiler/src/Language/Mimsa/Types/Error/InterpreterError.hs
… compiler/src/Language/Mimsa/Types/Error/ResolverError.hs
… compiler/src/Language/Mimsa/Types/Error/TypeError.hs
… compiler/src/Language/Mimsa/Types/Interpreter/Stack.hs
… compiler/src/Language/Mimsa/Types/Modules/Module.hs
… compiler/src/Language/Mimsa/Types/Project/Project.hs
… compiler/src/Language/Mimsa/Types/Project/SaveProject.hs
… compiler/src/Language/Mimsa/Types/Project/VersionedMap.hs