Thanks to https://kodimensional.dev/space-leak#use-strict-containers

Before:

```


benchmarking build stdlib/allFns
time                 203.0 ms   (197.9 ms .. 207.2 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 199.8 ms   (198.5 ms .. 201.2 ms)
std dev              1.814 ms   (1.310 ms .. 2.374 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking compilation/compile either fmap test
time                 10.86 ms   (10.80 ms .. 10.92 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 10.99 ms   (10.95 ms .. 11.04 ms)
std dev              124.8 μs   (93.54 μs .. 171.3 μs)

benchmarking evaluate/evaluate big looping thing
time                 110.8 ms   (108.5 ms .. 113.6 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 107.7 ms   (106.7 ms .. 109.4 ms)
std dev              1.970 ms   (1.454 ms .. 2.431 ms)

benchmarking evaluate/evaluate parsing
time                 76.45 ms   (75.64 ms .. 77.25 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 75.29 ms   (74.88 ms .. 75.69 ms)
std dev              743.5 μs   (598.4 μs .. 954.0 μs)

benchmarking evaluate/evaluate parsing 2
time                 307.1 ms   (304.0 ms .. 309.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 309.4 ms   (308.2 ms .. 310.5 ms)
std dev              1.321 ms   (777.8 μs .. 2.039 ms)
variance introduced by outliers: 16% (moderately inflated)
```

After:
```
benchmarking build stdlib/allFns
time                 184.6 ms   (183.1 ms .. 185.9 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 187.4 ms   (186.3 ms .. 189.5 ms)
std dev              2.102 ms   (563.5 μs .. 3.137 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking compilation/compile either fmap test
time                 9.968 ms   (9.930 ms .. 10.01 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 10.05 ms   (10.02 ms .. 10.10 ms)
std dev              82.83 μs   (56.34 μs .. 138.1 μs)

benchmarking evaluate/evaluate big looping thing
time                 97.33 ms   (96.58 ms .. 98.55 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 97.44 ms   (97.07 ms .. 97.82 ms)
std dev              610.7 μs   (490.5 μs .. 806.1 μs)

benchmarking evaluate/evaluate parsing
time                 69.64 ms   (68.95 ms .. 70.23 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 70.24 ms   (69.94 ms .. 70.52 ms)
std dev              516.4 μs   (390.8 μs .. 696.4 μs)

benchmarking evaluate/evaluate parsing 2
time                 282.5 ms   (281.0 ms .. 285.1 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 284.5 ms   (283.5 ms .. 285.1 ms)
std dev              1.110 ms   (620.2 μs .. 1.498 ms)
variance introduced by outliers: 16% (moderately inflated)
```